django_api/movies_admin/requirements.txt
asgiref==3.7.2
Django==4.2.7
django-debug-toolbar==4.2.0
django-split-settings==1.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
sqlparse==0.4.4
typing_extensions==4.8.0
